+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Time Block changelog
 
+## 0.0.2 -- not released
+
+- Working out support for time block view with dragging
+  - Using d3 to handle all of the time formatting and parsing
+
 ## 0.0.1 -- 2022-05-21 23:04:27
 
 - Initial version that supports a task list and basic list